oaba: BatchParamters interface/impl/integ-test/controller
BatchParameters
 * Removed setId(long) method
 * Removed RemoteException from the throws clause of all methods
 * Removed inheritance from EJBObject
 * Changed threshold fields from Long to long
   Knock-on
   + TransitivityOABAServiceBean

BatchParametersBean
 * Removed setId(long) method
 * Removed RemoteException from the throws clause of all methods
 * Removed inheritance from EntityBean
 * Added implements BatchParameters, Serializable
 * Added JPA annotations
 * Added methods for equals(Object), hashCode() and toString()

cmit
 * First working version of BatchParametersBeanIT
 * First working version of BatchParametersBeanController
 * Added BatchParametersBean to persistence.xml
 * Added table CMT_OABA_BATCH_PARAMS to DDL

+++ 2.7.x/choicemaker-cm/choicemaker-j2ee/com.choicemaker.cm.transitivity.server/src/main/java/com/choicemaker/cm/transitivity/server/impl/TransitivityOABAServiceBean.java
@@ -122,8 +122,8 @@ public class TransitivityOABAServiceBean implements SessionBean {
 			data.staging = batchParams.getStageRs();
 			data.stageModelName = batchParams.getStageModel();
 			data.masterModelName = batchParams.getMasterModel();
-			data.low = batchParams.getLowThreshold().floatValue();
-			data.high = batchParams.getHighThreshold().floatValue();
+			data.low = batchParams.getLowThreshold();
+			data.high = batchParams.getHighThreshold();
 			data.runTransitivity = false; //this means it's not a continuation from OABA.
 
 			sendToTransitivity (data);
